Add link to full API reference
Docs2.0 needs a link to the endpoint reference that currently still lives in Apiary. Adding this document is a workaround: it will be used to create a menu item in the lefthand sidebar whose content will (for now) just have a link to the Apiary docs.

NOTE the url it points to will point to Docs2.0 as soon as it's deployed: must be updated asap.